Move CompetitionPage into own component
