+++ src/components/CompetitionPage/RulesPage.tsx
@@ -1,0 +1,17 @@
+import {FC} from 'react'
+
+import {Markdown} from '@/components/StaticSites/Markdown'
+import {Competition} from '@/types/api/generated/competition'
+import {BannerContainer} from '@/utils/BannerContainer'
+import {formatDate} from '@/utils/formatDate'
+
+type RulesPageProps = Pick<Competition, 'rules' | 'upcoming_or_current_event' | 'name'>
+
+export const RulesPage: FC<RulesPageProps> = ({name, rules, upcoming_or_current_event}) => {
+  const {setBannerText} = BannerContainer.useContainer()
+
+  const startDate = formatDate(upcoming_or_current_event?.start)
+  setBannerText(upcoming_or_current_event ? `${name} sa bude konať  ${startDate}` : '')
+
+  return <Markdown content={rules ?? ''} />
+}
