+++ src/components/CompetitionPage/competition.module.scss
@@ -1,7 +1,6 @@
  .mainText{
     >p{
       padding: 20px;
-      width: 1000px;
     }
   }
