+++ src/components/CompetitionPage/competition.module.scss.d.ts
